requirements/demo.txt
django-devserver>=0.8,<0.9
